test: move spec.ts files to test directory
